feat: create service and repositories

+++ src/main/java/com/simplifiedpicpay/domain/repositories/TransactionRepository.java
@@ -4,6 +4,7 @@ import com.simplifiedpicpay.domain.transaction.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
-public interface TransactionRepository extends JpaRepository<Transaction, BigDecimal> {
+public interface TransactionRepository extends JpaRepository<Transaction, BigInteger> {
 }
